Move cookie consent to its own component
