+++ src/App.tsx
@@ -1,7 +1,5 @@
 import React, { Suspense, useContext, useEffect } from 'react'
 import CssBaseline from '@mui/material/CssBaseline'
-import CookieConsent from 'react-cookie-consent'
-import Cookies from 'js-cookie'
 import { createTheme, ThemeProvider } from '@mui/material/styles'
 import './index.css'
 import { Error } from './components/Error'
@@ -11,17 +9,14 @@ import {
   Route,
   createRoutesFromElements,
 } from 'react-router-dom'
-// this allows immer to work with Map and Set
-import { enableMapSet } from 'immer'
 import { MessagePanel } from './components/Messages'
 import appConfig from './assets/config.json'
 import { KeycloakContext } from './init/keycloak'
 import { useCredentialStore } from './store/CredentialStore'
 import { RedirectPanel } from './components/RedirectPanel'
 import ErrorBoundary from './components/ErrorBoundary'
+import { CookieConsentWidget } from './components/CookieConsent'
 // import { initHistoryClearing } from './store/hooks/useUrlNavigation/url-manager'
-
-enableMapSet()
 
 const AppShell = React.lazy(() => import('./components/AppShell'))
 const WorkspaceEditor = React.lazy(
@@ -98,51 +93,13 @@ export const App = (): React.ReactElement => {
     // initHistoryClearing()
   }, [])
 
-  const removeAllCookies = () => {
-    const allCookies = Cookies.get()
-    Object.keys(allCookies).forEach((cookieName) => {
-      Cookies.remove(cookieName, { path: '/' })
-    })
-  }
-
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
       <ErrorBoundary>
         <RouterProvider router={router} />
-        <CookieConsent
-          location="bottom"
-          buttonText="Accept"
-          declineButtonText="Decline"
-          enableDeclineButton
-          setDeclineCookie={false}
-          flipButtons
-          onDecline={removeAllCookies}
-          cookieName="cytoscapeWebCookieConsent"
-          style={{ background: '#4F4F4F' }}
-          buttonStyle={{
-            backgroundColor: '#0073B0',
-            color: '#ffffff',
-            fontSize: '13px',
-          }}
-          declineButtonStyle={{
-            color: '#ffffff',
-            background: '#6c757d',
-            fontSize: '13px',
-          }}
-          expires={150}
-        >
-          This site uses cookies to support Cytoscape Web’s network
-          visualization tools and improve your experience. By accepting, you
-          consent to our data practices.{' '}
-          <a
-            href="https://github.com/cytoscape/cytoscape-web/blob/development/privacy-policy.md"
-            style={{ color: '#e0e0e0' }}
-          >
-            Learn more
-          </a>
-        </CookieConsent>
       </ErrorBoundary>
+      <CookieConsentWidget />
     </ThemeProvider>
   )
 }
